abstract linked experiments for feature
